// src/interfaces/home/projectCardHome.ts
export interface ProjectCard {
  image: string;
  id?: number;
  slug?: string;
  status?: string;
  bedrooms?: number;
  bathrooms?: number;
  location: string;
  name: string;
  price: string;
  bed?: string;
  baths?: string;
  area?: string;
  construccion?: string;
  accentColor?: string;
  href?: string;
  longitude?: string;
  latitude?: string;
  region?: string;
  commune?: string;
  type?: string;
}